Working on proxy logic

+++ src/main/java/me/vasylkov/steamparser/parsing/model/Listing.java
@@ -3,6 +3,7 @@ package me.vasylkov.steamparser.parsing.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,7 +14,7 @@ public class Listing
     private String hashName;
     private double price;
     private List<Sticker> stickers;
-    private double floatValue;
+    private BigDecimal floatValue;
     private int pattern;
     private String imgUrl;
